+++ src/internals/getNonce.ts
@@ -1,10 +1,10 @@
-import { TsonNonce } from "../types.js";
+import { TsonNonce } from "../sync/syncTypes.js";
 
 const randomString = () => Math.random().toString(36).slice(2);
 
 export type GetNonce = () => TsonNonce;
 
-export const getNonce: GetNonce =
+export const getDefaultNonce: GetNonce =
 	typeof crypto === "object" && typeof crypto.randomUUID === "function"
 		? () => crypto.randomUUID() as TsonNonce
 		: () =>